fix(BA-5803): move metric types to repository layer to fix layer inversion
Metric data types (ContainerMetricOptionalLabel, ContainerMetricResponseInfo,
ContainerMetricResult, MetricResultValue) were defined in the service layer
but imported by the repository layer, violating the layer dependency rule.
Move them to repositories/metric/types.py and re-export from services for
backward compatibility. Also remove leftover empty processors/ directory.
