+++ src/ai/backend/manager/repositories/metric/types.py
@@ -5,8 +5,66 @@ from typing import Self, cast
 from uuid import UUID
 
 from ai.backend.common.clients.prometheus.types import MetricValue, ValueType
-from ai.backend.common.dto.clients.prometheus.response import PrometheusResponse
+from ai.backend.common.dto.clients.prometheus.response import (
+    MetricResponseInfo,
+    PrometheusResponse,
+)
+from ai.backend.common.exception import InvalidAPIParameters
 from ai.backend.common.types import KernelId
+
+
+@dataclass
+class ContainerMetricResponseInfo:
+    value_type: str
+    container_metric_name: str | None
+    agent_id: str | None
+    instance: str | None
+    job: str | None
+    kernel_id: str | None
+    owner_project_id: str | None
+    owner_user_id: str | None
+    session_id: str | None
+
+    @classmethod
+    def from_metric_response_info(cls, info: MetricResponseInfo) -> Self:
+        if info.value_type is None:
+            raise InvalidAPIParameters(
+                f"Missing required label 'value_type' for container metric (metric={info.name!r})"
+            )
+        return cls(
+            value_type=info.value_type,
+            container_metric_name=info.container_metric_name,
+            agent_id=info.agent_id,
+            instance=info.instance,
+            job=info.job,
+            kernel_id=info.kernel_id,
+            owner_project_id=info.owner_project_id,
+            owner_user_id=info.owner_user_id,
+            session_id=info.session_id,
+        )
+
+
+@dataclass
+class MetricResultValue:
+    timestamp: float
+    value: str
+
+
+@dataclass
+class ContainerMetricOptionalLabel:
+    value_type: ValueType
+
+    agent_id: str | None = None
+    kernel_id: UUID | None = None
+    session_id: UUID | None = None
+    user_id: UUID | None = None
+    project_id: UUID | None = None
+
+
+@dataclass
+class ContainerMetricResult:
+    metric: ContainerMetricResponseInfo
+    values: list[MetricResultValue]
 
 
 @dataclass(frozen=True)
